requirements.txt
dexter-cqa==1.0.9
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
